pyproject.toml
[project]
name = "2026-cognee-dify-hackathon"
version = "0.1.0"
description = ""
authors = [
    {name = "Dokeun Oh",email = "ohdoking@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "cognee (>=0.5.8,<0.6.0)",
    "asyncio (>=4.0.0,<5.0.0)",
    "requests (>=2.33.1,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "streamlit-mic-recorder (>=0.0.8,<0.0.9)",
    "streamlit (>=1.56.0,<2.0.0)",
    "openai (>=2.31.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
